# app/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ECLI API of OpenJustice",
        "description": "This is the documentation of ECLI Api of OpenJustice.be",
        "contact": {
            "name": "Team of OpenJustice",
            "email": "team@openjustice.be"
        },
        "license": {
            "name": "GNU General Public License v3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-ecli.openjustice.lltl.be/api/v1"
        }
    ],
    "paths": {
        "/ECLI/BE": {
            "get": {
                "tags": [
                    "category"
                ],
                "summary": "Get list of courts organised by categories",
                "description": "Get list of Courts with statistics organised by categories",
                "operationId": "Category",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/ECLI/BE/{court_acronym}": {
            "get": {
                "tags": [
                    "court"
                ],
                "summary": "Get Court information",
                "description": "Get Court",
                "operationId": "GetCourtInfo",
                "parameters": [
                    {
                        "name": "court_acronym",
                        "in": "path",
                        "description": "Court acronym",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/ECLI/BE/{court_acronym}/docsFilter": {
            "post": {
                "tags": [
                    "ECLI"
                ],
                "summary": "Filter documents of a court",
                "description": "Filter documents of a court",
                "operationId": "App\\Http\\Controllers\\Api\\DocumentController::docsFilter",
                "parameters": [
                    {
                        "name": "court_acronym",
                        "in": "path",
                        "description": "Court acronym",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "RSCE"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/filterDocs"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/ecli/post": {
            "post": {
                "tags": [
                    "ECLI"
                ],
                "summary": "Post new document",
                "description": "Post new document",
                "operationId": "App\\Http\\Controllers\\Api\\ECLIController::post",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/statistics": {
            "get": {
                "tags": [
                    "stats"
                ],
                "summary": "Get list of Stats",
                "description": "Get list of Stats",
                "operationId": "Stats",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/utus": {
            "get": {
                "tags": [
                    "utu"
                ],
                "summary": "Get list of Utus",
                "description": "Get list of Utus",
                "operationId": "utus",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "filterDocs": {
                "properties": {
                    "lang[]": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "french"
                        }
                    },
                    "year[]": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "2001"
                        }
                    },
                    "type[]": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "AVIS"
                        }
                    }
                },
                "type": "object"
            },
            "DocumentMinimalResource": {
                "title": "filterDocs",
                "description": "Document minumal resource",
                "properties": {
                    "identifier": {
                        "title": "identifier",
                        "type": "string",
                        "example": "20200526.HBOJ"
                    },
                    "type": {
                        "title": "type",
                        "type": "string",
                        "example": "DEC"
                    },
                    "type_identifier": {
                        "title": "type_identifier",
                        "description": "Concatenation of type, '.' and identifier",
                        "type": "string",
                        "example": "DEC.20200526.HBOJ"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at_diff": {
                        "title": "updated_at_diff",
                        "description": "updated_at field diff for humans",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "200": {
                "description": "Success"
            },
            "403": {
                "description": "Forbidden"
            },
            "404": {
                "description": "Resource Not Found"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "ECLI",
            "description": "Everything about ECLI"
        },
        {
            "name": "court",
            "description": "Everything about Court"
        },
        {
            "name": "category",
            "description": "Everything about category of courts"
        },
        {
            "name": "stats",
            "description": "Everything about category of stats"
        },
        {
            "name": "utus",
            "description": "Everything about Utus"
        }
    ]
}